+++ cookbook/src/app/Recipe/recipe-form/recipe-form.component.css
@@ -1,0 +1,9 @@
+.full-width {
+    width: 100%;
+}
+
+.recipeForm {
+    min-width: 150px;
+    max-width: 500px;
+    width: 100%;
+}
