--- cookbook/src/app/recipe-form/recipe-form.component.html
@@ -1,1 +1,0 @@
-<app-file-upload (fileEvent)="addImage($event)"></app-file-upload>
